requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
pymongo==4.8.0
python-dotenv==1.0.1
cryptography==43.0.1